feat(benchmarks): migrate all nine tasks to the row contract
Structural hooks emit gate and scored rows via one check() helper and end
exit 0; implement-feature's hand-rolled test staging becomes a hidden
tests/ overlay (symlinked baseline gate + five scored feature checks);
judge templates render {{GRADE_RESULT}}.
