+++ benchmarks/fit-wiki/tasks/cli-fix/hooks/invariants.sh
@@ -1,23 +1,26 @@
 #!/bin/sh
 set -u
-FAIL=0
-assert() {
-  id="$1"; shift
+# Emit one check row per probe; the rows carry the verdict and the exit code
+# stays script health only.
+row() {
+  name="$1"; extra="$2"; shift 2
   if "$@" >/dev/null 2>&1; then
-    echo "{\"id\":\"$id\",\"verdict\":\"pass\"}" >&"$RESULTS_FD"
+    printf '{"test":"%s","pass":true%s}\n' "$name" "$extra" >&"$RESULTS_FD"
   else
-    echo "{\"id\":\"$id\",\"verdict\":\"fail\"}" >&"$RESULTS_FD"
-    FAIL=1
+    printf '{"test":"%s","pass":false%s}\n' "$name" "$extra" >&"$RESULTS_FD"
   fi
 }
 
 # The fix must resolve the audit findings AND leave the seeded content
 # standing. `test -f` only proves the file still exists — a fix that guts
 # the body (or truncates it to satisfy the audit markers) would still pass.
-# Assert the seeded sections survive so the "intact" invariants actually
-# measure integrity instead of mere existence.
-assert audit-passes   sh -c 'cd "$1" && npx fit-wiki audit --today 2026-05-24' _ "$AGENT_CWD"
-assert summary-intact  sh -c 'f="$1/wiki/staff-engineer.md"; grep -q "## Current Focus" "$f" && grep -q "## Open Blockers" "$f"' _ "$AGENT_CWD"
-assert memory-intact   grep -q "## Cross-Cutting Priorities" "$AGENT_CWD/wiki/MEMORY.md"
+# The "intact" probes are anti-tamper gates so integrity failures can never
+# be traded for partial credit; the audit outcome is the scored check.
+row audit-passes '' \
+  sh -c 'cd "$1" && npx fit-wiki audit --today 2026-05-24' _ "$AGENT_CWD"
+row summary-intact ',"gate":true' \
+  sh -c 'f="$1/wiki/staff-engineer.md"; grep -q "## Current Focus" "$f" && grep -q "## Open Blockers" "$f"' _ "$AGENT_CWD"
+row memory-intact ',"gate":true' \
+  grep -q "## Cross-Cutting Priorities" "$AGENT_CWD/wiki/MEMORY.md"
 
-[ "$FAIL" = 0 ] && exit 0 || exit 1
+exit 0
