+++ benchmarks/kata-skills/tasks/coordinate-finding/hooks/invariants.sh
@@ -1,30 +1,31 @@
 #!/bin/sh
 set -u
 TRACKER="$AGENT_CWD/.tracker"
-FAIL=0
-assert() { fit-trace assert "$@" >&"$RESULTS_FD" || FAIL=1; }
+check() { fit-trace assert "$@" >&"$RESULTS_FD" || true; }
 
 # Resolve the single issue and change the loop should have produced. The ids are
 # caller-supplied slugs, so glob the tracker store rather than assume a name.
+# Empty globs fall back to a never-present path so the gates still emit rows.
 ISSUE=$(ls "$TRACKER"/issues/*.md 2>/dev/null | head -1)
 CHANGE=$(ls "$TRACKER"/changes/*.md 2>/dev/null | head -1)
+ISSUE="${ISSUE:-$TRACKER/issues/absent.md}"
+CHANGE="${CHANGE:-$TRACKER/changes/absent.md}"
 
-# An issue file must exist; without it nothing downstream can be asserted.
-assert issue-present --exists "$ISSUE"
-# A change file must exist before its envelope can be checked.
-assert change-present --exists "$CHANGE"
-[ "$FAIL" = 1 ] && exit 1
+# An issue and a change must exist; the linkage checks read both. Early exit
+# after a failing dependency gate — the gate row already carries the failure.
+fit-trace assert issue-present --gate --exists "$ISSUE" >&"$RESULTS_FD" || exit 0
+fit-trace assert change-present --gate --exists "$CHANGE" >&"$RESULTS_FD" || exit 0
 
 # The change links back to the issue (by its slug id, the filename stem).
 ISSUE_ID=$(basename "$ISSUE" .md)
-assert change-links-issue --grep "$ISSUE_ID" "$CHANGE" \
+check change-links-issue --grep "$ISSUE_ID" "$CHANGE" \
   --message "change does not link back to the issue"
 # The change reached merged state. Tolerate quoted front-matter values
 # (state: "merged"). Case-insensitive and multiline (RegExp /im/).
-assert change-merged --grep 'state:\s*["'"'"']?merged' "$CHANGE" \
+check change-merged --grep 'state:\s*["'"'"']?merged' "$CHANGE" \
   --message "change is not state: merged"
 # A trusted approval was recorded on the change (non-empty approval field).
-assert change-approved --grep 'approval:\s*\S' "$CHANGE" \
+check change-approved --grep 'approval:\s*\S' "$CHANGE" \
   --message "no approval recorded on the change"
 
-[ "$FAIL" = 0 ] && exit 0 || exit 1
+exit 0
