+++ benchmarks/kata-skills/tasks/design-feature/hooks/invariants.sh
@@ -1,24 +1,23 @@
 #!/bin/sh
 set -u
 DESIGN="$AGENT_CWD/specs/042-todo-filter/design-a.md"
-FAIL=0
-assert() { fit-trace assert "$@" >&"$RESULTS_FD" || FAIL=1; }
+check() { fit-trace assert "$@" >&"$RESULTS_FD" || true; }
 
-assert file-present --exists "$DESIGN"
-[ "$FAIL" = 1 ] && exit 1
+# The line count below reads the file, so the presence gate is a dependency:
+# early exit after a failing gate — the gate row already carries the failure.
+fit-trace assert file-present --gate --exists "$DESIGN" >&"$RESULTS_FD" || exit 0
 
-# Design § over-200-lines is a Blocker (kata-review delta).
+# Design § over-200-lines is a Blocker (kata-review delta) — a gate.
 LINES=$(wc -l < "$DESIGN")
 if [ "$LINES" -lt 200 ]; then
-  echo "{\"test\":\"under-200-lines\",\"pass\":true}" >&"$RESULTS_FD"
+  printf '%s\n' '{"test":"under-200-lines","pass":true,"gate":true}' >&"$RESULTS_FD"
 else
-  echo "{\"test\":\"under-200-lines\",\"pass\":false,\"message\":\"$LINES lines\"}" >&"$RESULTS_FD"
-  FAIL=1
+  printf '%s\n' "{\"test\":\"under-200-lines\",\"pass\":false,\"gate\":true,\"message\":\"$LINES lines\"}" >&"$RESULTS_FD"
 fi
 
 # A decisions heading at any depth ("## Decisions", "### Key Decision"), and any
 # of the common ways a rejected option gets named. Case-insensitive (RegExp /im/).
-assert has-decisions  --grep '^#{2,6}[ \t]+.*Decision' "$DESIGN"
-assert names-tradeoff --grep 'reject|alternative|instead of|rather than|trade.?off' "$DESIGN"
+check has-decisions  --grep '^#{2,6}[ \t]+.*Decision' "$DESIGN"
+check names-tradeoff --grep 'reject|alternative|instead of|rather than|trade.?off' "$DESIGN"
 
-[ "$FAIL" = 0 ] && exit 0 || exit 1
+exit 0
