+++ benchmarks/kata-skills/tasks/implement-feature/judge.task.md
@@ -1,7 +1,7 @@
-Invariants result:
+Grade result:
 
 ```json
-{{INVARIANTS_RESULT}}
+{{GRADE_RESULT}}
 ```
 
 The agent was given these instructions:
@@ -12,8 +12,9 @@ Agent trace at `{{AGENT_TRACE_PATH}}`. The implemented app is under
 `{{TASK_DIR}}/app/`; the approved spec, design, and plan are in
 `{{TASK_DIR}}/specs/042-todo-filter/`.
 
-The invariants already ran the hidden tests, so behaviour is graded
-mechanically. Your job is **discipline**: did the agent implement the plan
+The harness already ran the hidden test suite and restored the working
+directory, so behaviour is graded mechanically and every file you see is the
+agent's own work. Your job is **discipline**: did the agent implement the plan
 faithfully with **no scope creep** — no refactors, renames, extra features, or
 changes to `add`/`done`/the stored shape beyond what `plan-a.md` describes? Read
 the modified files under `app/` and the trace. Did it follow the instructions
