+++ benchmarks/kata-skills/tasks/implement-feature/tests/app/test/filter-no-match.test.js
@@ -1,0 +1,8 @@
+import { test } from "node:test";
+import assert from "node:assert/strict";
+
+import { store, sample } from "./feature-helpers.js";
+
+test("filterTodos returns nothing when no match", () => {
+  assert.deepEqual(store.filterTodos(sample, "zzz"), []);
+});
