+++ benchmarks/kata-skills/tasks/implement-feature/tests/app/test/todo.gate.test.js
@@ -1,0 +1,1 @@
+../../../../../workdir/app/test/todo.test.js
